fix clippy

+++ pallets/proxy/src/benchmarking.rs
@@ -270,7 +270,7 @@ benchmarks! {
     )
     verify {
         let proxies = EVMProxies::<T>::get(evm_address);
-        let last_proxy = proxies.get(0 as usize).ok_or("last pure proxy not found")?;
+        let last_proxy = proxies.get(0_usize).ok_or("last pure proxy not found")?;
         ensure!(Proxies::<T>::contains_key(last_proxy), "pure proxy not created");
         assert_last_event::<T>(Event::PureCreated {
             pure: last_proxy.clone(),
